refactor: MovieItem 컴포넌트에서 포스터 URL 생성 로직 분리 및 이미지 경로 상수 추가 정리
