+++ src/constants/imagePaths.ts
@@ -1,16 +1,3 @@
-// TMDB API 이미지 기본 경로
-const TMDB_IMAGE_BASE_URL = "https://image.tmdb.org/t/p";
-
-// 이미지 크기 옵션
-const POSTER_SIZES = {
-  SMALL: "w154",
-  MEDIUM: "w185",
-  LARGE: "w342",
-  X_LARGE: "w500",
-  MOVIE_DETAIL: "w440_and_h660_face",
-  ORIGINAL: "original",
-};
-
 export const POSTER_PATH = Object.freeze({
   DEFAULT: "./default-poster.svg",
   LOADING: "./loading-poster.svg",
@@ -18,19 +5,16 @@ export const POSTER_PATH = Object.freeze({
 });
 
 export const ICON_PATH = Object.freeze({
-  STAR: "./star_empty.png",
+  STAR_EMPTY: "./star_empty.png",
+  STAR_FILLED: "./star_filled.png",
   SEARCH: "./search.svg",
+  MODAL_CLOSE: "./modal_button_close.svg",
+});
+
+export const LOGO_PATH = Object.freeze({
+  LOGO: "./woowacourse_logo.png",
 });
 
 export const IMAGE_PATH = Object.freeze({
   EMPTY_PLANET: "./empty-planet.svg",
 });
-
-// 포스터 URL 생성 함수
-export const getPosterUrl = (
-  posterPath: string | null,
-  size = POSTER_SIZES.MOVIE_DETAIL
-): string => {
-  if (!posterPath) return POSTER_PATH.DEFAULT;
-  return `${TMDB_IMAGE_BASE_URL}/${size}${posterPath}`;
-};
